requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
python-telegram-bot==22.1
sniffio==1.3.1
typing_extensions==4.13.2
yarl==1.20.0
